backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
python-dotenv==1.0.1
pdf2image==1.17.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
pytesseract==0.3.13
requests==2.32.3
pydantic==2.9.0
PyJWT==2.9.0
openai>=2.24.0
